Phase 2: Extract import/export functionality
Refactoring changes:
- Created constants.py to centralize all configuration constants
- Created import_export/ package for import/export functionality
- Extracted ImportWorker class to import_export/import_worker.py
- Created CSVExporter class in import_export/csv_exporter.py
- Updated AstroFileManager.py to use new modules
- Removed duplicate ImportWorker class definition (281 lines)
- Simplified CSV export methods by delegating to CSVExporter

Benefits:
- Better separation of concerns
- Avoids circular import dependencies
- Reduces main file from 3,085 → 2,815 lines
- Import/export logic now modular and testable

All functionality preserved and syntax verified.
